src/requirements.txt
-i https://pypi.org/simple
joblib==1.3.2
structlog==24.1.0
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.12.0
threadpoolctl==3.2.0
requests==2.31.0
pandas==2.0.2
apache-airflow[sentry]